requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
async-timeout==4.0.3
attrs==25.3.0
backoff==2.2.1
blinker==1.9.0
boto3==1.38.18
botocore==1.38.18
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
cloudpickle==3.1.1
cmudict==1.0.32
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dacite==1.6.0
dagshub==0.5.9
dagshub-annotation-converter==0.1.9
databricks-sdk==0.53.0
dataclasses-json==0.6.7
datasets==3.6.0
Deprecated==1.2.18
dill==0.3.8
distro==1.9.0
docker==7.1.0
entrypoints==0.4
evaluate==0.4.3
exceptiongroup==1.3.0
faiss-cpu==1.11.0
fastapi==0.115.12
filelock==3.18.0
Flask==3.1.1
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.1
gql==3.5.2
graphene==3.4.3
graphql-core==3.2.4
graphql-relay==3.2.0
greenlet==3.2.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.27.2
huggingface-hub==0.31.4
idna==3.10
importlib_metadata==7.2.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.1.16
langchain-community==0.0.33
langchain-core==0.1.53
langchain-openai==0.0.8
langchain-text-splitters==0.0.2
langsmith==0.1.147
limits==5.2.0
load-dotenv==0.1.0
lxml==5.4.0
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mdurl==0.1.2
mlflow==2.22.0
mlflow-skinny==2.22.0
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
mypy_extensions==1.1.0
networkx==3.4.2
numpy==1.26.4
openai==1.79.0
opentelemetry-api==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
orjson==3.10.18
packaging==23.2
pandas==2.2.3
pathvalidate==3.2.3
pillow==11.2.1
propcache==0.3.1
protobuf==5.29.4
pyarrow==15.0.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.2
pywin32==310
PyYAML==6.0.2
querystring-parser==1.2.4
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9.1
s3transfer==0.12.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
semver==3.0.4
six==1.17.0
slowapi==0.1.9
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
starlette==0.46.2
sympy==1.14.0
tenacity==8.5.0
textstat==0.7.7
threadpoolctl==3.6.0
tiktoken==0.6.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.52.3
treelib==1.7.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
waitress==3.0.2
watchfiles==1.0.5
Werkzeug==3.1.3
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
